Check the REDCap Log

Retrieve and analyze the REDCap log to monitor recent activity or filter for specific users and time ranges.

Usage

check_REDCap_log(
  DB,
  last = 24,
  user = "",
  units = "hours",
  begin_time = "",
  clean = T,
  record = ""
)

Arguments

DB

A validated `DB` object containing REDCap project data and settings. Generated using `load_DB()` or `setup_DB()`

last

Numeric. Specifies the time range paired with `units` to filter the log. Default is `24`.

user

Character. Optional user filter to retrieve log entries associated with a specific username. Default is an empty string (`""`), which includes all users.

units

Character. Units paired with `last` to define the time range. Options are `"mins"`, `"hours"`, `"days"`. Default is `"hours"`.

begin_time

Character. A specific timestamp (e.g., "2023-07-11 13:15:06") to define the start of the log retrieval. Overrides `last` and `units` if provided. Default is an empty string (`""`).

clean

Logical. If `TRUE`, cleans the API data and adds extra summary columns to the log. Default is `TRUE`.

record

Character. Optional record filter for retrieving log entries associated with a specific record. Default is an empty string (`""`).

Value

A `data.frame` containing the cleaned and summarized REDCap log.

Details

This function queries the REDCap project log to retrieve recent activity based on the specified time range, user, or starting timestamp. It optionally cleans the raw API data and adds summary columns to enhance log readability and analysis.

See also