Check the REDCap log

Retrieves raw or labeled data from a REDCap database based on the provided parameters. The function connects to a REDCap project using the REDCap API, fetches the raw or labeled data (depending on the `labelled` parameter), and returns the data as a `data.frame`. Optionally, specific records can be retrieved.

Usage

get_REDCap_raw_data(DB, labelled = F, records = NULL, batch_size = 1000)

Arguments

DB

A validated `DB` object containing REDCap project data and settings. Generated using load_DB or setup_DB

labelled

Logical. If `TRUE`, the function retrieves labeled data; if `FALSE`, it retrieves raw data. Default is `FALSE`.

records

Optional character vector of record IDs to retrieve. If `NULL`, all records are fetched.

batch_size

Numeric. The batch size for the retrieval process. Default is `1000`.

Value

A `data.frame` containing the raw or labeled REDCap data.

Details

This function uses the redcap_read function to fetch data from a REDCap project. The function allows for batch retrieval, and the `labelled` argument determines whether the data is returned with raw values or with labels. By default, the function retrieves raw data. It also supports filtering for specific records if provided.

See also