Set a Valid REDCap API Token an R Session

Prompts the user to input a valid REDCap API token and stores it as an environment variable for the current R session. This function validates the token before storing it. It is strongly discouraged to include API tokens directly within R scripts. Instead of using this function you should consider setting your token within your REnviron file which can be edited with `usethis::edit_r_environ()`

Usage

set_REDCap_token(DB, ask = T)

Arguments

DB

A validated list object that contains all metadata, data, logs, transformations, and more for one REDCap project.

ask

Logical (TRUE/FALSE). If TRUE, asks the user for confirmation before overwriting an existing valid token. Default is `TRUE`.

Value

Invisible. A message is printed to confirm the token is successfully set.

Details

If a valid token already exists in the R session, the function notifies the user and asks whether they want to replace it. The user is guided to provide a new token through the console prompt. The token is validated internally and stored using `Sys.setenv()`.

See also