Setup for DB including token

Setup for DB including token

setup_RosyREDCap(
  short_name,
  dir_path,
  token_name,
  redcap_base,
  force = F,
  merge_form_name = "merged",
  validate = T,
  use_csv = F
)

Arguments

short_name

character name as a shortcut

dir_path

character file path of the directory

token_name

character string of what the token is called when using Sys.setenv and Sys.getenv

redcap_base

character of the base REDCap link, ex. https://redcap.miami.edu

force

logical for force blank load vs last save

merge_form_name

name of merged non-repeating to be used in package

validate

logical for validation

DB

a list object slightly modified version of RosyDB and contains all metadata, data, transformations, and more

Value

DB