Summarize REDCap Database

Summarizes the REDCap database (`DB` object) by filtering and generating a summary list.

Usage

summarize_DB(
  DB,
  with_links = T,
  deidentify = T,
  clean = T,
  drop_blanks = T,
  include_metadata = T,
  annotate_metadata = T,
  include_record_summary = T,
  include_users = T,
  include_log = T,
  separate = F,
  force = F
)

Arguments

DB

A validated `DB` object containing REDCap project data and settings. Generated using load_DB or setup_DB

with_links

Logical (TRUE/FALSE). If TRUE, includes links in the summary. Default is `TRUE`.

deidentify

Logical (TRUE/FALSE). If TRUE, deidentifies the summary data. Default is `TRUE`.

clean

Logical (TRUE/FALSE). If TRUE, cleans the summary data. Default is `TRUE`.

drop_blanks

Logical (TRUE/FALSE). If TRUE, drops blank entries from the summary. Default is `TRUE`.

include_metadata

Logical (TRUE/FALSE). If TRUE, includes metadata in the summary. Default is `TRUE`.

annotate_metadata

Logical (TRUE/FALSE). If TRUE, annotates metadata in the summary. Default is `TRUE`.

include_record_summary

Logical (TRUE/FALSE). If TRUE, includes a summary of records in the summary. Default is `TRUE`.

include_users

Logical (TRUE/FALSE). If TRUE, includes user information in the summary. Default is `TRUE`.

include_log

Logical (TRUE/FALSE). If TRUE, includes logs in the summary. Default is `TRUE`.

separate

Logical (TRUE/FALSE). If TRUE, separates the summary into different sections. Default is `FALSE`.

force

Logical (TRUE/FALSE). If TRUE, forces the summary generation even if there are issues. Default is `FALSE`.

Value

List. Returns a list containing the summarized data, including records, metadata, users, logs, and any other specified data.

Details

This function filters the REDCap database based on the provided parameters and generates a summary list. The summary can include metadata, record summaries, user information, and logs. The function also supports deidentification and cleaning of the data.

See also