Shows DB in the env

Shows DB in the env

update_RosyREDCap(
  DB,
  force = F,
  day_of_log = 10,
  labelled = T,
  get_files = F,
  original_file_names = F,
  entire_log = F,
  metadata_only = F,
  ask_about_overwrites = T,
  save_to_dir = T,
  batch_size = 2000
)

Arguments

DB

DB from load_RosyREDCap or setup_DB

force

logical for force a fresh update

day_of_log

numbers of days to be checked in the log

labelled

logical for whether or not to return raw or labelled REDCap. Default is TRUE.

get_files

logical for whether or not to get files from redcap.

original_file_names

logical for whether or not to use original file names.

Value

messages for confirmation