Shows DB in the env

Shows DB in the env

Usage

drop_redcap_dir(
  DB,
  records,
  smart = T,
  include_metadata = T,
  include_other = T,
  with_links = T,
  forms,
  merge_non_repeating = T,
  separate = F,
  str_trunc_length = 32000,
  file_name
)

Arguments

records

character vector of records you want dropped to your directory

smart

logical for whether to only save when data is new

include_metadata

logical for whether to only include redcap and not metadata

include_other

logical for whether to only include redcap and not metadata

with_links

optional logical for including links in excel sheets

forms

optional character vector for only selected forms

merge_non_repeating

optional logical for merging non-repeating instruments

str_trunc_length

optional integer for truncation

file_name

optional character string for adding to the front of file names

allow_mod

logical for whether non-instrument names are allowed

deidentify

logical for deidentification

dir_other

optional character string of another file path where the files should be saved

Value

messages for confirmation