Setup a DB Object for a REDCap Project

Initializes a RosyREDCap `DB` object from one REDCap project and if a directory is chosen it will structure that directory for later imports and exports.

Usage

setup_RosyREDCap(
  short_name,
  dir_path,
  redcap_base,
  force = F,
  merge_form_name = "merged",
  use_csv = F,
  auto_check_token = T
)

Arguments

short_name

Character string with no spaces or symbols. A short unique name to use as a shortcut for the project such PSDB.

dir_path

Character. File path of the directory where project data is stored.

redcap_base

Character. The base URL of the REDCap server, e.g., "https://redcap.miami.edu".

force

Logical. If `TRUE`, forces a blank load instead of loading the last saved state. Default is `FALSE`.

merge_form_name

Character. Name of the merged non-repeating form to use in the package. Default is `"merged"`.

use_csv

Logical. If `TRUE`, uses CSV files instead of other storage formats. Default is `FALSE`.

auto_check_token

Logical. If `TRUE`, automatically tests the REDCap API token after setup. Default is `TRUE`.

DB

A validated list object that contains all metadata, data, logs, transformations, and more for one REDCap project.

Value

A `DB` object configured for the REDCap project.

Details

This function creates or loads a `DB` object containing all the necessary elements for working with a REDCap project. It supports integration with REDCap servers, handles metadata configuration, and optionally validates the REDCap API token on setup.

The function allows customization of project-specific settings such as the base REDCap URL, file paths, and merge form names. If the `force` argument is `TRUE`, the function skips loading a previously saved project and initializes a blank setup, which may be helpful if you are getting errors.

See also